Fix macos warnings about output format casting.

+++ output.cpp
@@ -418,7 +418,7 @@ namespace marxan {
     void displayTimePassed(chrono::high_resolution_clock::time_point start)
     {
         auto end = chrono::high_resolution_clock::now();
-        uint64_t itemp = chrono::duration_cast<std::chrono::seconds>(end - start).count();
+        int itemp = static_cast<int>(chrono::duration_cast<std::chrono::seconds>(end - start).count());
 
         printf("Time passed so far is ");
         if (itemp >= 60 * 60)
